requirements.txt
fastapi
uvicorn
gunicorn

sqlalchemy
psycopg2-binary

python-dotenv

passlib[bcrypt]
python-jose

email-validator
pydantic
pydantic[email]
pydantic-settings

alembic

python-multipart

requests
